fix(diagnostics): restore narrow OTEL disabled admission
