+++ extensions/diagnostics-otel/src/service-propagation.ts
@@ -1,0 +1,125 @@
+import {
+  context,
+  createContextKey,
+  diag,
+  propagation,
+  ROOT_CONTEXT,
+  type Context,
+  type TextMapGetter,
+  type TextMapPropagator,
+  type TextMapSetter,
+} from "@opentelemetry/api";
+import { AsyncLocalStorageContextManager } from "@opentelemetry/context-async-hooks";
+import {
+  CompositePropagator,
+  getStringListFromEnv,
+  W3CBaggagePropagator,
+  W3CTraceContextPropagator,
+} from "@opentelemetry/core";
+import { B3InjectEncoding, B3Propagator } from "@opentelemetry/propagator-b3";
+import { JaegerPropagator } from "@opentelemetry/propagator-jaeger";
+
+const DEFAULT_PROPAGATORS = ["tracecontext", "baggage"];
+const CONTEXT_OWNER_KEY = createContextKey("openclaw.disabled-sdk.context-owner");
+const PROPAGATOR_OWNER_KEY = createContextKey("openclaw.disabled-sdk.propagator-owner");
+
+class OwnedContextManager extends AsyncLocalStorageContextManager {
+  constructor(private readonly owner: object) {
+    super();
+  }
+
+  override active(): Context {
+    return super.active().setValue(CONTEXT_OWNER_KEY, this.owner);
+  }
+}
+
+class OwnedPropagator implements TextMapPropagator {
+  constructor(
+    private readonly delegate: TextMapPropagator,
+    private readonly owner: object,
+  ) {}
+
+  inject(carrierContext: Context, carrier: unknown, setter: TextMapSetter): void {
+    const probe = carrierContext.getValue(PROPAGATOR_OWNER_KEY);
+    if (probe && typeof probe === "object") {
+      (probe as { owner?: object }).owner = this.owner;
+      return;
+    }
+    this.delegate.inject(carrierContext, carrier, setter);
+  }
+
+  extract(carrierContext: Context, carrier: unknown, getter: TextMapGetter): Context {
+    return this.delegate.extract(carrierContext, carrier, getter);
+  }
+
+  fields(): string[] {
+    return this.delegate.fields();
+  }
+}
+
+function ownsGlobalPropagator(owner: object): boolean {
+  const probe: { owner?: object } = {};
+  propagation.inject(ROOT_CONTEXT.setValue(PROPAGATOR_OWNER_KEY, probe), {}, { set() {} });
+  return probe.owner === owner;
+}
+
+function createConfiguredPropagator(): TextMapPropagator | null {
+  const names = (getStringListFromEnv("OTEL_PROPAGATORS") ?? DEFAULT_PROPAGATORS).map((name) =>
+    name.toLowerCase(),
+  );
+  if (names.includes("none")) {
+    return null;
+  }
+  const propagators = [...new Set(names)].flatMap((name): TextMapPropagator[] => {
+    switch (name) {
+      case "tracecontext":
+        return [new W3CTraceContextPropagator()];
+      case "baggage":
+        return [new W3CBaggagePropagator()];
+      case "b3":
+        return [new B3Propagator()];
+      case "b3multi":
+        return [new B3Propagator({ injectEncoding: B3InjectEncoding.MULTI_HEADER })];
+      case "jaeger":
+        diag.warn(
+          'The Jaeger propagator is deprecated and will be removed in a future release. Use the W3C TraceContext propagator ("tracecontext") instead.',
+        );
+        return [new JaegerPropagator()];
+      default:
+        diag.warn(`Propagator "${name}" requested through environment variable is unavailable.`);
+        return [];
+    }
+  });
+  if (propagators.length === 0) {
+    return null;
+  }
+  return propagators.length === 1 ? propagators[0]! : new CompositePropagator({ propagators });
+}
+
+export function registerDisabledSdkRuntime(): (() => void) | null {
+  const owner = {};
+  const contextManager = new OwnedContextManager(owner).enable();
+  const ownsContext = context.setGlobalContextManager(contextManager);
+  if (!ownsContext) {
+    contextManager.disable();
+  }
+  const propagator = createConfiguredPropagator();
+  const ownsPropagation = propagator
+    ? propagation.setGlobalPropagator(new OwnedPropagator(propagator, owner))
+    : false;
+  if (!ownsContext && !ownsPropagation) {
+    return null;
+  }
+  // Ownership can change after registration. Probe the current public API
+  // delegate before removing a global so a later host SDK remains intact.
+  return () => {
+    if (ownsPropagation && ownsGlobalPropagator(owner)) {
+      propagation.disable();
+    }
+    if (ownsContext && context.active().getValue(CONTEXT_OWNER_KEY) === owner) {
+      context.disable();
+    } else if (ownsContext) {
+      contextManager.disable();
+    }
+  };
+}
